feat(entity): update phong entity

+++ backend/SoTroOnline/src/main/java/com/so_tro_online/app/QuanLyPhong/src/main/java/com/so_tro_online/quan_ly_phong/dto/RoomRequest.java
@@ -14,6 +14,6 @@ public class RoomRequest {
     private BigDecimal chieuDai;
     private BigDecimal chieuRong;
     private String vatDung;
-    private Float giaThueCoBan;
+    private BigDecimal giaThueCoBan;
     private TrangThai trangThai;
 }
